Refactor attribute imports to seperate file.
Otherwise it somehow messes with builtins on Python 2.7.
